% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.r
\name{drop_nulls}
\alias{drop_nulls}
\title{Drop NULL values from a list}
\usage{
drop_nulls(x)
}
\arguments{
\item{x}{A list from which to remove NULL values.}
}
\value{
A list with NULL values removed.
}
\description{
Drop NULL values from a list
}
