% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.r
\name{get}
\alias{get}
\title{Make a GET request to Paddle API}
\usage{
get(link)
}
\arguments{
\item{link}{The API endpoint URL.}
}
\value{
Parsed JSON response from the API.
}
\description{
Make a GET request to Paddle API
}
