% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paddle_mode.R
\name{get_paddle_key}
\alias{get_paddle_key}
\title{Get Paddle API Key Based on Current Mode}
\usage{
get_paddle_key()
}
\value{
A character string representing the Paddle API key.
}
\description{
Resolves the Paddle API key based on environment variables and the selected mode (live or sandbox).
}
\examples{
\dontshow{if (paddle_has_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_paddle_key()  # returns token for current mode
\dontshow{\}) # examplesIf}
}
