% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.r
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value.}

\item{b}{Second value used if `a` is NULL.}
}
\value{
Either `a` or `b`.
}
\description{
Returns `a` if not NULL, otherwise returns `b`.
}
