% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{paddle_cancel_subscription}
\alias{paddle_cancel_subscription}
\title{Cancel a Paddle subscription}
\usage{
paddle_cancel_subscription(subscription_id, effective_from = NULL)
}
\arguments{
\item{subscription_id}{Character. Required. Paddle subscription ID, e.g. "sub_abc123".}

\item{effective_from}{Character or NULL. Optional. One of `"next_billing_period"` or `"immediately"`. Defaults to `"next_billing_period"`.}
}
\value{
A list with the updated subscription entity and metadata.
}
\description{
Cancels a subscription using its ID. Defaults to cancel at next billing period unless `effective_from` is set to "immediately".
}
