% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{paddle_list_products}
\alias{paddle_list_products}
\title{Retrieve Paddle Products}
\usage{
paddle_list_products(
  id = NULL,
  status = NULL,
  tax_category = NULL,
  type = NULL,
  include = NULL,
  after = NULL,
  order_by = NULL,
  per_page = NULL
)
}
\arguments{
\item{id}{Character vector of product IDs to match. Optional.}

\item{status}{Character vector of statuses to filter by (e.g., \code{"active"}, \code{"archived"}). Optional.}

\item{tax_category}{Character vector of tax categories to filter. One of \code{"digital-goods"}, \verb{"ebooks}, \code{"implementation-services"}, \code{"professional-services"}, \code{"saas"}, \code{"software-programming-services"}, \code{"standard"}, \code{"training-services"}, and \code{"website-hosting"}. Optional.}

\item{type}{Character. Type of product (one of \code{"standard"} and \code{"custom"}). Optional.}

\item{include}{Character vector of related entities - default \code{NULL} or \code{"prices"}. Optional.}

\item{after}{Character. Return entities after the specified Paddle ID when working with paginated endpoints.. Optional.}

\item{order_by}{Character. Ordering string (e.g., \code{"id[ASC]"}). Valid fields for ordering: \code{"created_at"}, \code{"custom_data"}, \code{"id"}, \code{"description"}, \code{"image_url"}, \code{"name"}, \code{"status"}, \code{"tax_category"}, and \code{"updated_at"}. Valid directions \code{"[ASC]"} and \code{"[DESC]"} Optional. Optional.}

\item{per_page}{Number of products per page (max 200). Optional, defaults to 50.}
}
\value{
A list containing product data and pagination metadata.
}
\description{
Fetches a paginated list of products from the Paddle API. By default, only active products are returned.
}
\examples{
\dontshow{if (paddle_has_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set_paddle_mode("sandbox")
result <- paddle_list_products()
\dontshow{\}) # examplesIf}
}
