% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{paddle_update_transaction}
\alias{paddle_update_transaction}
\title{Update a Transaction}
\usage{
paddle_update_transaction(
  transaction_id,
  status = NULL,
  customer_id = NULL,
  address_id = NULL,
  business_id = NULL,
  custom_data = NULL,
  currency_code = NULL,
  collection_mode = NULL,
  discount_id = NULL,
  billing_details = NULL,
  payment_terms = NULL,
  billing_period = NULL,
  items = NULL,
  checkout = NULL,
  include = NULL
)
}
\arguments{
\item{transaction_id}{Character. Required. The ID of the transaction (e.g. `"txn_abc123"`).}

\item{status}{Character. Optional. You may set a transaction to `"billed"` or `"canceled"`. Billed transactions cannot be changed.}

\item{customer_id}{Character. Optional. Paddle customer ID.}

\item{address_id}{Character. Optional. Requires `customer_id`.}

\item{business_id}{Character. Optional. Requires `customer_id`.}

\item{custom_data}{Named list. Optional structured metadata.}

\item{currency_code}{Character. Optional. Must be `"USD"`, `"EUR"`, or `"GBP"` if `collection_mode = "manual"`.}

\item{collection_mode}{Character. Optional. Must be `"automatic"` or `"manual"`.}

\item{discount_id}{Character. Optional. Paddle discount ID.}

\item{billing_details}{List. Required if `collection_mode = "manual"`. May include:
- `payment_terms` → list with `interval` (Character, must be one of `"day"`, `"week"`, `"month"`, `"year"`) and `frequency` (integer)
- `enable_checkout` (logical)
- `purchase_order_number` (Character)
- `additional_information` (Character or null)}

\item{billing_period}{List with `starts_at` and `ends_at` (RFC 3339 format). Optional.}

\item{items}{List of items. Optional. Each object must contain either a `price_id` or a `price object`, and a `quantity`. Please review API documentation for more details.}

\item{checkout}{List with `url`. Optional.}

\item{include}{Character vector. Related entities to include. Must be one of `"address"`, `"adjustments"`, `"adjustments_totals"`, `"available_payment_methods"`, `"business"`, `"customer"`, `"discount"`.}
}
\value{
A list with updated transaction data and metadata.
}
\description{
Updates a transaction by its ID. Only transactions with status `draft` or `ready` can be updated.
}
