% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.r
\name{post_excl_body}
\alias{post_excl_body}
\title{Make a POST request to Paddle API without body}
\usage{
post_excl_body(link)
}
\arguments{
\item{link}{The API endpoint URL.}

\item{body}{The body of the request, typically a list to be converted to JSON.}
}
\value{
Parsed JSON response from the API.
}
\description{
Make a POST request to Paddle API without body
}
