% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.r
\name{update}
\alias{update}
\title{Make a DELETE request to Paddle API}
\usage{
update(link, body)
}
\arguments{
\item{link}{The API endpoint URL.}

\item{body}{The body of the request, typically a list to be converted to JSON.}
}
\value{
Parsed JSON response from the API.
}
\description{
Make a DELETE request to Paddle API
}
