% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paddle_mode.R
\name{get_paddle_mode}
\alias{get_paddle_mode}
\title{Get Paddle Mode
This function retrieves the current mode set for Paddle API requests.}
\usage{
get_paddle_mode()
}
\value{
The current mode for Paddle API requests, either "live" or "sandbox".
}
\description{
Get Paddle Mode
This function retrieves the current mode set for Paddle API requests.
}
