% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{paddle_create_price}
\alias{paddle_create_price}
\title{Create a Paddle Price}
\usage{
paddle_create_price(
  description,
  product_id,
  unit_price,
  type = NULL,
  name = NULL,
  billing_cycle = NULL,
  trial_period = NULL,
  tax_mode = NULL,
  unit_price_overrides = NULL,
  quantity = NULL,
  custom_data = NULL
)
}
\arguments{
\item{description}{Internal description for your team. Required.}

\item{product_id}{ID of the product this price belongs to. Required.}

\item{unit_price}{A list with `amount` (string, lowest denomination, e.g. for 10 USD write 1000 (lowest denomination = cents)) and `currency_code` (3-letter ISO). Required.}

\item{type}{Type of item (one of `"standard"` and `"custom"`). Optional. Defaults to `"standard"`.}

\item{name}{Name of this price (shown at checkout/invoices). Optional.}

\item{billing_cycle}{List with `frequency` and `interval` (one of `"day"`, `"week"`, `"month"`, `"year"`). Optional. Omit for one-time pricing.}

\item{trial_period}{List with `frequency` and `interval` (one of `"day"`, `"week"`, `"month"`, `"year"`). Optional. Requires `billing_cycle`.}

\item{tax_mode}{One of `"account_setting"`, `"external"`, `"internal"`. Optional.}

\item{unit_price_overrides}{A list of overrides with `country_codes` (Supported two-letter ISO 3166-1 alpha-2 country code) and `unit_price` (same list as in unit_price parameter). Optional.}

\item{quantity}{List with `minimum` and `maximum` quantity limits. Optional.}

\item{custom_data}{Named list of custom metadata. Optional.}
}
\value{
A list representing the created price entity and metadata.
}
\description{
Creates a new price for a product. The price defines how much customers pay, how often, and under what terms.
}
\examples{
\dontrun{
set_paddle_mode("sandbox")
paddle_create_price(
 description = "Standard monthly subscription",
 product_id = "pro_01jwp8mvatdtncfca757r5wb9f",
 unit_price = list(
 amount = "1000",  # 10.00 USD
 currency_code = "USD"
 ),
)
}
}
