% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paddle_mode.R
\name{paddle_has_token}
\alias{paddle_has_token}
\title{Check if Paddle API Token is Available
This function checks whether a Paddle API key is available in the environment,
either as `PADDLE_KEY` or `PADDLE_KEY_LIVE` / `PADDLE_KEY_SANDBOX`.}
\usage{
paddle_has_token()
}
\value{
Logical `TRUE` if a token is found, `FALSE` otherwise.
}
\description{
Check if Paddle API Token is Available
This function checks whether a Paddle API key is available in the environment,
either as `PADDLE_KEY` or `PADDLE_KEY_LIVE` / `PADDLE_KEY_SANDBOX`.
}
\examples{
paddle_has_token()
}
\keyword{internal}
